// src/utils/enums.ts
export enum ClientEventType {
	CALL_STARTED = 'CALL_STARTED',
	CALL_ENDED = 'CALL_ENDED',
	CLIENT_JOINED = 'CLIENT_JOINED',
	CLIENT_LEFT = 'CLIENT_LEFT',
	PEER_CONNECTION_OPENED = 'PEER_CONNECTION_OPENED',
	PEER_CONNECTION_CLOSED = 'PEER_CONNECTION_CLOSED',
	MEDIA_TRACK_ADDED = 'MEDIA_TRACK_ADDED',
	MEDIA_TRACK_REMOVED = 'MEDIA_TRACK_REMOVED',
	MEDIA_TRACK_RESUMED = 'MEDIA_TRACK_RESUMED',
	MEDIA_TRACK_MUTED = 'MEDIA_TRACK_MUTED',
	MEDIA_TRACK_UNMUTED = 'MEDIA_TRACK_UNMUTED',
	ICE_GATHERING_STATE_CHANGED = 'ICE_GATHERING_STATE_CHANGED',
	PEER_CONNECTION_STATE_CHANGED = 'PEER_CONNECTION_STATE_CHANGED',
	ICE_CONNECTION_STATE_CHANGED = 'ICE_CONNECTION_STATE_CHANGED',
	DATA_CHANNEL_OPEN = 'DATA_CHANNEL_OPEN',
	DATA_CHANNEL_CLOSED = 'DATA_CHANNEL_CLOSED',
	DATA_CHANNEL_ERROR = 'DATA_CHANNEL_ERROR',
	NEGOTIATION_NEEDED = 'NEGOTIATION_NEEDED',
	SIGNALING_STATE_CHANGE = 'SIGNALING_STATE_CHANGE',
	ICE_GATHERING_STATE_CHANGE = 'ICE_GATHERING_STATE_CHANGE',
	ICE_CONNECTION_STATE_CHANGE = 'ICE_CONNECTION_STATE_CHANGE',
	ICE_CANDIDATE = 'ICE_CANDIDATE',
	
	CLIENT_ISSUE = 'CLIENT_ISSUE',

	// mediasoup events
	PRODUCER_ADDED = 'PRODUCER_ADDED',
	PRODUCER_REMOVED = 'PRODUCER_REMOVED',
	CONSUMER_ADDED = 'CONSUMER_ADDED',
	CONSUMER_REMOVED = 'CONSUMER_REMOVED',
	DATA_PRODUCER_CREATED = 'DATA_PRODUCER_CREATED',
	DATA_PRODUCER_CLOSED = 'DATA_PRODUCER_CLOSED',
	DATA_CONSUMER_CREATED = 'DATA_CONSUMER_CREATED',
	DATA_CONSUMER_CLOSED = 'DATA_CONSUMER_CLOSED',
}